requirements.txt
requests==2.31.0
reportlab==4.0.9
Pillow==10.2.0
